add stars

+++ src/shaders/dot-fragment.glsl
@@ -1,8 +1,7 @@
-precision mediump float;
+precision highp float;
 
 uniform vec3 uColor;
 
 void main() {
-//  gl_FragColor = vec4(uColor.xyz, 1.0);
  gl_FragColor = vec4(1.0, 0.0, 1.0, 1.0);
 }
